Tuned Color Sort
Changed the ring detection to use distance from the optical sensor and use color to identify if we want to keep or toss the ring. Still not 100% reliable. Currently at a point where it is semi consistent while driving and standing still. Since this is going to be used most during auto, probably best to focus on tuning while moving.